run: Change text base of Spike platform to 0x60000000
Change text base of Spike platform from 0x80000000 to 0x60000000.

When loading a branch target address as a 32-bit word with 'lw'
instruction with text-base = 0x80000000, the result will be
sign-extended to 0xFFFFFFFFxxxxxxxx, which is not a valid text
address.

For example, in making original slice with Gem5:
  lw a5, 0(a5) : MemRead :  D=0x000000000001a354 A=0x7ffea8d5cf10

in running reconstructed slice on Spike with text-base = 0x80000000:
  0x0000000080000034 (0x0062b023) | sd t1, 0(t0) |  mem 0x0000000000071d90 0x800035c000000000
  ...
  0x00000000800021d0 (0x0007a783) | lw a5, 0(a5) |  x15 0xffffffff800035c0 mem 0x0000000000071d94
  0x00000000800021d4 (0x000780e7) | jalr a5 |  x1 0x00000000800021d8
  (Spike halts)
the final 'jalr' would not jump to 0x800035c0 as we expected.

While running on Spike with text-base = 0x60000000:
  0x00000000600021d0 (0x0007a783) | lw a5, 0(a5) |  x15 0x00000000600035c0 mem 0x0000000000071d94
  0x00000000600021d4 (0x000780e7) | jalr a5 |  x1 0x00000000600021d8
  0x00000000600035c0 (0x0409e993) | ori s3, s3, 64 |  x19 0x0000000000000041
  ...
No side effect is introduced by the sign-extension of 'lw', and the
following 'jalr' would do a correct jumping.